% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addTblDesc.R
\name{addTblDesc}
\alias{addTblDesc}
\title{add to table descriptiion}
\usage{
addTblDesc(df, db = NULL, overwrite = F)
}
\arguments{
\item{df}{a dataframe that adheres to the standard of a spatial resolution table}

\item{db}{The database to add terms to.  Alternatively you can set this as an option termDB with options(termDB = "myDB.sqlite") and leave the this parameter out}

\item{overwrite}{If TRUE, overwrite the underlying table.  This may be a very bad idea.  So be careful!}

\item{dcheck}{If FALSE, no duplicate checking is done!  This may be a bad idea!}
}
\description{
This will add to a table of table descriptions
}
\details{
a terms table is made up of the following headings, please consult the database definitional document for details.
}

