% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkNames}
\alias{checkNames}
\title{Check headers}
\usage{
checkNames(nlist, plist)
}
\arguments{
\item{nlist}{a list of header names to check against}

\item{plist}{possible headers}
}
\description{
Just a helper function to make sure that headers are the same
}

