% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDB}
\alias{createDB}
\title{create DB}
\usage{
createDB(path)
}
\arguments{
\item{path}{the path and name of your database. E.g. "test.sqlite" will create a database of that name in your current working directory}
}
\description{
This will create an sqlite3 database for you.
}
\details{
You'll just need to call this once to create your database
}

