% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createID.R
\name{createID}
\alias{createID}
\title{assign a uniue ID}
\usage{
createID(numid, tbl, db)
}
\arguments{
\item{numid}{The number of new id's to return}

\item{tbl}{the table to get id's from}

\item{db}{the database name}
}
\value{
a vector of true and false indicating which rows of potential input should be added to the database
}
\description{
Will assign a unique ID to a new set of rows by picking up at the end of the id's in the table
}

