% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createTbl}
\alias{createTbl}
\title{create a database table}
\usage{
createTbl(tbl, db, namelist, types)
}
\arguments{
\item{tbl}{the name of the table to create}

\item{db}{the database to create the table in}

\item{namelist}{A named list of table headings}

\item{types}{A list of types that each column should be, e.g. TEXT or INT}
}
\description{
This will create a table for you from named list of terms and types
}
\examples{
\dontrun{
 tbl <- "mytable"
 db <- "test.db"
 namelist <- c("id","desc")
 types <- c("INT","TEXT")
 createTbl(tbl,db,namelist,types)
}
}

