% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractTables}
\alias{extractTables}
\title{extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet}
\usage{
extractTables(tab)
}
\description{
extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet
}

