% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractTablesIS.R
\name{extractTablesIS}
\alias{extractTablesIS}
\title{extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet}
\usage{
extractTablesIS(tab)
}
\description{
extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet
}

