% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inputDataPubIS}
\alias{inputDataPubIS}
\title{input an *IS data publication workbook into the db}
\usage{
inputDataPubIS(datapub, temporal, horizontal, vertical, db)
}
\description{
Based on a standard template, input a data publication workbook
}

