% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/termDefLookUp.R
\name{termDefLookUp}
\alias{termDefLookUp}
\title{checks terms in an input workbook against existing terms in the database
returns a data frame of termName, description given in the workbook, and
description in the database if the descriptions don't match}
\usage{
termDefLookUp(wk)
}
\description{
checks terms in an input workbook against existing terms in the database
returns a data frame of termName, description given in the workbook, and
description in the database if the descriptions don't match
}

