% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/termLookUp.R
\name{termLookUp}
\alias{termLookUp}
\title{output all uses of a given term, in a table of dpID, termID, tableID, termName, DPName}
\usage{
termLookUp(x)
}
\description{
output all uses of a given term, in a table of dpID, termID, tableID, termName, DPName
}

