% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkNames.R
\name{checkNames}
\alias{checkNames}
\title{Check headers}
\usage{
checkNames(nlist, plist)
}
\arguments{
\item{nlist}{a list of header names to check against}

\item{plist}{possible headers}
}
\description{
Just a helper function to make sure that headers are the same
}

