% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extractTables.R
\name{extractTables}
\alias{extractTables}
\title{extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet}
\usage{
extractTables(tab)
}
\description{
extract DPD, TempRes, and SpRes tables from input ingest or publication worksheet
}

