% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inputDataPub.R
\name{inputDataPub}
\alias{inputDataPub}
\title{input a data publication workbook into the db}
\usage{
inputDataPub(datapub, temporal, space, db)
}
\description{
Based on a standard template, input a data publication workbook
}

