% Generated by roxygen2 (4.0.1): do not edit by hand
\name{termLookUp}
\alias{termLookUp}
\title{output all uses of a given term, in a table of dpID, termID, tableID, termName, DPName}
\usage{
termLookUp(x)
}
\description{
output all uses of a given term, in a table of dpID, termID, tableID, termName, DPName
}

