% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testTbl.R
\name{testTbl}
\alias{testTbl}
\title{Test table}
\usage{
testTbl(tbl, db)
}
\arguments{
\item{tbl}{the table name you want to check}

\item{db}{the database to find the table in}
}
\value{
TRUE if exists, FALSE if not
}
\description{
simple helper function to test whether or not a table exists
}

